pdf_text_converter/requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0

# PDF processing
pdf2image>=1.16.0
# NOTE: Also requires poppler-utils (install separately)

# Google Cloud Vision API
google-cloud-vision>=3.0.0

# Table detection backends
transformers>=4.25.0
torch>=1.12.0
paddleocr>=2.6.0
paddlepaddle>=2.4.0

# Visualization
matplotlib>=3.5.0

# Progress bars (optional)
tqdm>=4.64.0